+++ liveauctions-backend/src/main/java/com/ute/auction/controller/seller/RegistrationProductController.java
@@ -26,53 +26,60 @@ public class RegistrationProductController {
 
     // Build API sorted asc starting price
     @GetMapping("/registration-products/{id}/sorted-price-asc")
-    public ResponseEntity<List<RegistrationProductDTO>> sortedAscByStartingPrice(@PathVariable("id") Long sellerId, 
-                                                                                @RequestParam("page") int page, 
-                                                                                @RequestParam(defaultValue = "10") int size) {
+    public ResponseEntity<List<RegistrationProductDTO>> sortedAscByStartingPrice(@PathVariable("id") int sellerId,
+            @RequestParam("page") int page,
+            @RequestParam(defaultValue = "10") int size) {
         List<RegistrationProductDTO> models = registrationProductService.sortedAscByStartingPrice(sellerId, page, size);
         return ResponseEntity.ok(models);
     }
 
     // Build API sorted desc starting price
     @GetMapping("/registration-products/{id}/sorted-price-desc")
-    public ResponseEntity<List<RegistrationProductDTO>> sortedDescByStartingPrice(@PathVariable("id") Long sellerId, 
-                                                                                @RequestParam("page") int page, 
-                                                                                @RequestParam(defaultValue = "10") int size) {
-        List<RegistrationProductDTO> models = registrationProductService.sortedDescByStartingPrice(sellerId, page, size);
+    public ResponseEntity<List<RegistrationProductDTO>> sortedDescByStartingPrice(@PathVariable("id") int sellerId,
+            @RequestParam("page") int page,
+            @RequestParam(defaultValue = "10") int size) {
+        List<RegistrationProductDTO> models = registrationProductService.sortedDescByStartingPrice(sellerId, page,
+                size);
         return ResponseEntity.ok(models);
     }
 
     // Build API sorted asc registration date
     @GetMapping("/registration-products/{id}/sorted-date-asc")
-    public ResponseEntity<List<RegistrationProductDTO>> sortedAscByRegistrationDate(@PathVariable("id") Long sellerId, 
-                                                                                    @RequestParam("page") int page, 
-                                                                                    @RequestParam(defaultValue = "10") int size) {
-        List<RegistrationProductDTO> models = registrationProductService.sortedAscByRegistrationDate(sellerId, page, size);
+    public ResponseEntity<List<RegistrationProductDTO>> sortedAscByRegistrationDate(@PathVariable("id") int sellerId,
+            @RequestParam("page") int page,
+            @RequestParam(defaultValue = "10") int size) {
+        List<RegistrationProductDTO> models = registrationProductService.sortedAscByRegistrationDate(sellerId, page,
+                size);
         return ResponseEntity.ok(models);
     }
 
     // Build API sorted desc registration date
     @GetMapping("/registration-products/{id}/sorted-date-desc")
-    public ResponseEntity<List<RegistrationProductDTO>> sortedDescByRegistrationDate(@PathVariable("id") Long sellerId, 
-                                                                                    @RequestParam("page") int page, 
-                                                                                    @RequestParam(defaultValue = "10") int size) {
-        List<RegistrationProductDTO> models = registrationProductService.sortedDescByRegistrationDate(sellerId, page, size);
+    public ResponseEntity<List<RegistrationProductDTO>> sortedDescByRegistrationDate(@PathVariable("id") int sellerId,
+            @RequestParam("page") int page,
+            @RequestParam(defaultValue = "10") int size) {
+        List<RegistrationProductDTO> models = registrationProductService.sortedDescByRegistrationDate(sellerId, page,
+                size);
         return ResponseEntity.ok(models);
-    }    
+    }
 
     // Build API get registration products by seller id
     @GetMapping("/registration-products/{id}")
-    public ResponseEntity<List<RegistrationProductDTO>> getRegistrationProductsBySellerId(@PathVariable("id") Long sellerId, 
-                                                                                        @RequestParam("page") int page, 
-                                                                                        @RequestParam(defaultValue = "10") int size) {
-        List<RegistrationProductDTO> registrationProducts = registrationProductService.getRegistrationProductsBySellerId(sellerId, page, size);
+    public ResponseEntity<List<RegistrationProductDTO>> getRegistrationProductsBySellerId(
+            @PathVariable("id") int sellerId,
+            @RequestParam("page") int page,
+            @RequestParam(defaultValue = "10") int size) {
+        List<RegistrationProductDTO> registrationProducts = registrationProductService
+                .getRegistrationProductsBySellerId(sellerId, page, size);
         return ResponseEntity.ok(registrationProducts);
     }
 
     // Build API register product
     @PostMapping("/register-product")
-    public ResponseEntity<RegistrationProductDTO> registerProduct(@RequestBody RegistrationProductDTO registrationProductDTO) {
-        RegistrationProductDTO savedRegistrationProduct = registrationProductService.registerProduct(registrationProductDTO);
+    public ResponseEntity<RegistrationProductDTO> registerProduct(
+            @RequestBody RegistrationProductDTO registrationProductDTO) {
+        RegistrationProductDTO savedRegistrationProduct = registrationProductService
+                .registerProduct(registrationProductDTO);
         return new ResponseEntity<>(savedRegistrationProduct, HttpStatus.OK);
     }
