+++ liveauctions-backend/src/main/java/com/ute/auction/service/IProductService.java
@@ -6,10 +6,10 @@ import com.ute.auction.dto.ProductDTO;
 
 public interface IProductService {
 
-    List<ProductDTO> getProductsBySellerId(Long sellerId, int page, int size);
+    List<ProductDTO> getProductsBySellerId(int sellerId, int page, int size);
 
-    List<ProductDTO> sortedAscByStartingPrice(Long sellerId, int page, int size);
+    List<ProductDTO> sortedAscByStartingPrice(int sellerId, int page, int size);
 
-    List<ProductDTO> sortedDescByStartingPrice(Long sellerId, int page, int size);
-    
+    List<ProductDTO> sortedDescByStartingPrice(int sellerId, int page, int size);
+
 }
