+++ liveauctions-backend/src/main/java/com/ute/auction/service/IRegistrationProductService.java
@@ -6,16 +6,16 @@ import com.ute.auction.dto.RegistrationProductDTO;
 
 public interface IRegistrationProductService {
 
-    List<RegistrationProductDTO> getRegistrationProductsBySellerId(Long sellerId, int page, int size);
+    List<RegistrationProductDTO> getRegistrationProductsBySellerId(int sellerId, int page, int size);
 
     RegistrationProductDTO registerProduct(RegistrationProductDTO regisProduct);
 
-    List<RegistrationProductDTO> sortedAscByStartingPrice(Long sellerId, int page, int size);
+    List<RegistrationProductDTO> sortedAscByStartingPrice(int sellerId, int page, int size);
 
-    List<RegistrationProductDTO> sortedDescByStartingPrice(Long sellerId, int page, int size);
+    List<RegistrationProductDTO> sortedDescByStartingPrice(int sellerId, int page, int size);
 
-    List<RegistrationProductDTO> sortedAscByRegistrationDate(Long sellerId, int page, int size);
+    List<RegistrationProductDTO> sortedAscByRegistrationDate(int sellerId, int page, int size);
 
-    List<RegistrationProductDTO> sortedDescByRegistrationDate(Long sellerId, int page, int size);
-    
+    List<RegistrationProductDTO> sortedDescByRegistrationDate(int sellerId, int page, int size);
+
 }
